\name{normwhn.test-package}
\alias{norwhn.test-package}
\alias{normwhn.test}
\docType{package}
\title{Normality and White Noise Testing}
\description{Multivariate and Univariate Normality Tests (See Doornik and Hansen (1994)). Also a Univariate Test for White Noise (See Lobato and Velasco (2004))}
\details{
\tabular{ll}{
Package: \tab normwhn.test\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2010-08-02\cr
License: \tab GPL-3\cr
}
}
\author{Peter Wickham
Maintainer: Peter Wickham <peterwickham@mac.com}
\references{
Doornik, J.A., and H. Hansen (1994). "An Omnibus Test for Univariate and Multivariate Normality", Working Paper, Nuffield College,
Oxford University, U.K.
Lobato, I., and C. Velasco (2004). "A Simple Test of Normality of Time Series", Econometric Theory, 20, pp. 671-689, Cambridge University Press.
Lobato, I., and C. Velasco (2004). "A Simple and General Test for White Noise", Econometric Society, Latin-America Meetings, Paper No. 112 
}
\keyword{ package }

